Issue in main algorithm, have it for certain ones, but not others.
Calling it a night for now, very close to a solution. It works for some values but seems to have issues with dealing with others. It might be best to test each method by itself in order to locate the issue. Its either the skeleton's generation fault or the parity calculation's fault or it could be both.

Anyways a quick summary of what I did to the code. I removed its dependency on being a main method and made it into a proper class. This makes it more flexible and allows quicker testing because the input can just be pasted into a temporary main. Also by having it be a class we can easily insert it into the GUI form. I then went through your code and noticed there were two main parts. The first part was finding out what the encoded frame should look like and then filling in that frame. I took the first part and created the generateEncodedFrameSkeleton method to handle that. The filling in of the frame can be done using a for loop because the pattern can be calculated by a function which makes it easy to pick out the bits we need to check. I changed the char arrays out for a int array because it means we can just add the bits we want to check and mod them by two to find out if they are even or odd. Finally I made the arrays instance variables because the frame is now a fully fledged object.

TODO:
Fix the algorithm that does the encoding.
Finish cleaning up and converting the code to an object styled class.
Also need to rename the class because it doesn't follow java's style "framing" => "Framing"