Changed "HammingCheck" -> "HammingDecode"
Just so the class is more descriptive and matches the HammingEncode class.

Stylistically it would be best just to have a Hamming class with a decode and encode function in it. It could be done easily, but I don't want to risk breaking stuff right now. Also the classes are descriptive enough and don't really cause confusion.